+++ src/stochastic_analyzer/Dockerfile
@@ -3,8 +3,8 @@ FROM python:3.13-slim
 WORKDIR /app
 
 COPY src/shared_functions/ src/shared_functions/
-COPY src/stochastic_analyzer/ src/stochastic_analyzer/
+COPY src/stochastic_analyzer_v2/ src/stochastic_analyzer_v2/
 
-RUN pip install --no-cache-dir ./src/stochastic_analyzer
+RUN pip install --no-cache-dir ./src/stochastic_analyzer_v2
 
 ENTRYPOINT ["gateway"]
