+++ src/stochastic_analyzer/setup.py
@@ -10,11 +10,9 @@ setup(
         "fastapi[standard] >= 0.128",
         "uvicorn",
         "pydantic >= 2.0.0",
-        "requests>=2.32",
-        "httpx",
+        "aiohttp>=3.13",
         "markdown-pdf>=1.13.1",
-        "markitdown[docx,pptx,xlsx,pdf]>=0.1",
-        "tokenizers>=0.20",
+        "markitdown[pptx, docx, xlsx, xls, pdf]>=0.1.5",
         f"shared-functions @ {shared_lib_path}",
     ]
 )
