--- src/stochastic_analyzer_v2/Dockerfile
@@ -1,10 +1,0 @@
-FROM python:3.13-slim
-
-WORKDIR /app
-
-COPY src/shared_functions/ src/shared_functions/
-COPY src/stochastic_analyzer_v2/ src/stochastic_analyzer_v2/
-
-RUN pip install --no-cache-dir ./src/stochastic_analyzer_v2
-
-ENTRYPOINT ["gateway"]
