--- src/stochastic_analyzer_v2/pyproject.toml
@@ -1,15 +1,0 @@
-[build-system]
-requires = ["setuptools>=77.0.3"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "stochastic-analyzer"
-version = "1.0.0"
-description = "Unified query gateway service"
-dynamic = ["dependencies"]
-
-[tool.setuptools]
-packages = ["gateway", "gateway.services"]
-
-[project.scripts]
-gateway = "gateway.main:start"
